core/rpc/client/user/http:  read and interpret error messages
